Move pkg operation/project/state -> operation/state
