+++ pkg/lib/operation/state/load/operation.go
